requirements.txt
boto3==1.4.6
botocore==1.6.3
click==6.7
docutils==0.14
jmespath==0.9.3
PyAudio==0.2.11
python-dateutil==2.6.1
python-dotenv==0.6.5
s3transfer==0.1.10
six==1.10.0
webrtcvad==2.0.10
